fix(components): adjust icns and username ui
