+++ src/lib/components/PrimaryNameMark.tsx
@@ -1,8 +1,8 @@
 import { CustomIcon } from "lib/components/icon";
 import { Tooltip } from "lib/components/Tooltip";
 
-export const PrimaryNameMark = () => (
+export const PrimaryNameMark = ({ boxSize = 3 }: { boxSize?: number }) => (
   <Tooltip label="Primary name">
-    <CustomIcon name="star-solid" color="accent.main" boxSize={3} m={0} />
+    <CustomIcon name="star-solid" color="accent.main" boxSize={boxSize} m={0} />
   </Tooltip>
 );
