+++ src/lib/layout/Header.tsx
@@ -6,7 +6,6 @@ import { FaucetBtn } from "lib/components/button";
 import { WalletSection } from "lib/components/Wallet";
 
 import { NetworkMenu } from "./NetworkMenu";
-// import Searchbar from "./Searchbar";
 import { SearchComponent } from "./SearchComponent";
 
 const Header = () => {
@@ -34,7 +33,6 @@ const Header = () => {
           _hover={{ cursor: "pointer", opacity: 0.85 }}
         />
       </AppLink>
-      {/* <Searchbar /> */}
       <SearchComponent />
       <Flex gap={4}>
         <FaucetBtn />
